Extract and reuse shared helper
