+++ packages/commands/src/query/contracts/show.ts
@@ -5,7 +5,7 @@ import { z } from "zod";
 
 import { DEFAULT_EVM_RPC_URL } from "../../../../../src/constants/api";
 import { contractsShowOptionsSchema } from "../../../../../src/schemas/commands/contracts";
-import { formatAddress } from "../../../../../utils/addressResolver";
+import { formatAddressForChain } from "../../../../../utils/addressResolver";
 import { fetchContracts } from "./list";
 
 type ContractsShowOptions = z.infer<typeof contractsShowOptionsSchema>;
@@ -15,29 +15,6 @@ interface ContractData {
   chainId: ethers.BigNumberish;
   contractType: string;
 }
-
-/**
- * For Sui chains (chain IDs 103 and 105), contract addresses are 64 bytes (128 hex chars)
- * representing two concatenated 32-byte values. Split them and display on separate lines.
- */
-const splitSuiCombinedAddress = (bytesHex: string): string => {
-  const hex = bytesHex.startsWith("0x") ? bytesHex.slice(2) : bytesHex;
-  if (hex.length !== 128) return formatAddress(bytesHex);
-  const partA = hex.slice(0, 64);
-  const partB = hex.slice(64, 128);
-  return `0x${partA}\n0x${partB}`;
-};
-
-const formatAddressForChain = (
-  bytesHex: string,
-  chainId: ethers.BigNumberish
-): string => {
-  const id = chainId.toString();
-  if (id === "103" || id === "105") {
-    return splitSuiCombinedAddress(bytesHex);
-  }
-  return formatAddress(bytesHex);
-};
 
 const findContractByChainId = (
   contracts: ContractData[],
